+++ HIRS_ProvisionerTPM2/include/Process.h
@@ -20,6 +20,9 @@ class Process {
  private:
     static const hirs::log::Logger LOGGER;
 
+    static const char* const kPgrepCommand;
+    static const int kMaxStatFileProcessNameLength;
+
     std::string executable;
 
     std::string arguments;
@@ -40,6 +43,8 @@ class Process {
                            std::string sourceFileName,
                            int sourceLineNumber,
                            const std::string& arguments = "");
+
+    static bool isRunning(const std::string& executable);
 };
 
 }  // namespace utils
