+++ HIRS_ProvisionerTPM2/test/Process_Test.cpp
@@ -10,6 +10,10 @@
 #include "log4cplus/configurator.h"
 #include "gtest/gtest.h"
 
+using hirs::utils::Process;
+using std::string;
+using std::stringstream;
+
 namespace {
 
 class ProcessTest : public :: testing::Test {
@@ -36,53 +40,73 @@ class ProcessTest : public :: testing::Test {
 };
 
 TEST_F(ProcessTest, ProcessWorks) {
-    hirs::utils::Process p("echo \"Hello World\"");
+    Process p("echo \"Hello World\"");
     int retVal = p.run();
     ASSERT_EQ(retVal, 0);
     ASSERT_EQ("Hello World\n", p.getOutputString());
 }
 
 TEST_F(ProcessTest, ProcessTwoArgConstructorWorks) {
-    hirs::utils::Process p("echo", "\"Hello World\"");
+    Process p("echo", "\"Hello World\"");
     int retVal = p.run();
     ASSERT_EQ(retVal, 0);
     ASSERT_EQ("Hello World\n", p.getOutputString());
 }
 
 TEST_F(ProcessTest, ProcessFailsWithNonZeroReturnValue) {
-    hirs::utils::Process p("ls", "isjlfidjsaij");
+    Process p("ls", "isjlfidjsaij");
     int retVal = p.run();
     ASSERT_EQ(retVal, 2);
 }
 
 TEST_F(ProcessTest, NonExistentProcessFailsWithNonZeroReturnValue) {
-    hirs::utils::Process p("isjlfidjsaij");
+    Process p("isjlfidjsaij");
     int retVal = p.run();
     ASSERT_EQ(retVal, 127);
 }
 
 TEST_F(ProcessTest, NonExistentProcessFailsAndGivesErrorMessage) {
-    hirs::utils::Process p("isjlfidjsaij", "ijijdfi");
-    std::stringstream expectedError;
+    Process p("isjlfidjsaij", "ijijdfi");
+    stringstream expectedError;
     expectedError << "Call to isjlfidjsaij returned 127" << std::endl
                   << "Is isjlfidjsaij in your path?" << std::endl;
-    std::string expectedErrorString(expectedError.str());
+    string expectedErrorString(expectedError.str());
 
-    std::stringstream errorStream;
+    stringstream errorStream;
     int retVal = p.run(errorStream);
     ASSERT_EQ(retVal, 127);
-    std::string receivedErrorString(errorStream.str());
+    string receivedErrorString(errorStream.str());
     ASSERT_EQ(receivedErrorString, expectedErrorString);
 }
 
 TEST_F(ProcessTest, SuccessfulProcessDoesNotProduceErrorMessage) {
-        hirs::utils::Process p("echo", "\"Hello World\"");
+    Process p("echo", "\"Hello World\"");
 
-    std::stringstream errorStream;
+    stringstream errorStream;
     int retVal = p.run(errorStream);
     ASSERT_EQ(retVal, 0);
-    std::string receivedErrorString(errorStream.str());
+    string receivedErrorString(errorStream.str());
     ASSERT_TRUE(receivedErrorString.empty());
+}
+
+TEST_F(ProcessTest, ProcessIsRunningSuccessful) {
+    ASSERT_TRUE(Process::isRunning("Process"));
+}
+
+TEST_F(ProcessTest, ProcessIsRunningSuccessfulPathBased) {
+    ASSERT_TRUE(Process::isRunning("/opt/Process"));
+}
+
+TEST_F(ProcessTest, ProcessIsRunningFalse) {
+    ASSERT_FALSE(Process::isRunning("foobar"));
+}
+
+TEST_F(ProcessTest, ProcessIsRunningEmptyStringReturnsFalse) {
+    ASSERT_FALSE(Process::isRunning(""));
+}
+
+TEST_F(ProcessTest, ProcessIsRunningPreventCommandHijack) {
+    ASSERT_FALSE(Process::isRunning("foobar; echo blarg"));
 }
 
 }  // namespace
